+++ tests/test.lunaris.index.js
@@ -186,4 +186,14 @@ describe('lunaris index', () => {
     should(index.invalidations._invalidations).be.ok();
     should(index.invalidations._invalidations).be.an.Object();
   });
+
+  it('debug should be defined', () => {
+    should(index.debug).be.ok();
+    should(index.debug).be.an.Object();
+    should(index.debug.isDebug).be.a.Boolean();
+    should(index.debug.config).be.an.Object();
+    should(index.debug.debug).be.a.Function();
+    should(index.debug.log).be.a.Function();
+    should(index.debug.NAMESPACES).be.an.Object();
+  });
 });